Initially migrated everything to a model, no testing yet.
